feat(entity, repository): Add Budget entity; Create repositories for each entity; Add comments
The Budget entity was created here and can be linked to a user as well as a category so a user can budget for a specific category. Many budgets can be linked to one user. Many budgets can be linked to one category. In addition, comments were added to each entity to better label and organize the properties and relationship mappings in the code.