Add bootstrap package with test data to test entity relationship mappings are working correctly. Update Transaction amount field data type from double to BigDecimal
To verify that the entity relationships were correctly mapped, I created a package named bootstrap with a file named TestDataLoader.java. In that file, I used a CommandLineRunner to create test data on start up to confirm that application can correctly connect to the database and create some data and persist it to the database.