feat(exception, service): update exeption package to have only one exception used for services; update UserService, BudgetService, and CategoryService to no longer use optional<T> data types
Before, I had a different exception file for each service, but the code in them was so similar, I consolidated it to be just one file. If in the future, unique custom exceptions are needed, then I can extend this single NotFoundException.java file that was created. Furthermore, I also updated the methods that returned optional<T> data types to just returning T types, and the methods will throw the exception if something is not found